Timer, Edit Game Dialog functionality
Completely refurbished the runplayer panel to regainfocus if there is a
change to any run, meaning it automatically finds the active run and
selects it again, also it correctly displays informationif you resort runs
in the Run Control panel.

Fully integrated changes to a specific run to both reflect in Run Control,
Active Run, And other panels

Made improvements to the horaro import plugin to write out meaningful
errors if there is a problem with the import or if the import has the
wrong format. For instance it will now debug that the last run of the
preesa marathon is missing the "category" and "platform" fields, but will
continue to import anyways.

Finalized the "Change Game Data" dialog