Big changes to the stopwatch code
Most of the code in the extension\stopwatch.js folder was copy/pasted
from a file from the SGDQ15 layouts, and most of it was unneeded; it was
designed for multiple timers when we only use one main timer anyway. I
cleaned it up so only what we need is now there and changed the
"stopwatches" replicant to "stopwatch" and it is now a single object and
not an array.