Small "workaround" on the Team Layout panel
If there are more than 1 teams (a team can just be 1 player in a race),
the panel is the same as before, but if there is only 1 team, the panel
changes to sorting of that team's members; useful for co-op runs with
more than 1 screen. I don't plan to keep it this way forever, but it
could help out. (Ignore the copy/pasted code!)